+++ compiler/startNode.json
@@ -1,0 +1,43 @@
+{
+    "id": "start",
+    "data": {
+      "id": "start",
+      "name": "start",
+      "type": "Output",
+      "label": "Start",
+      "inputs": {},
+      "outputs": {
+        "startPointer": ""
+      },
+      "category": "Miscellaneous",
+      "selected": false,
+      "baseClasses": [
+        "xMessage"
+      ],
+      "description": "Start pointer Points to node  ",
+      "inputParams": [],
+      "inputAnchors": [],
+      "outputAnchors": [
+        {
+          "id": "start-output-startPointer-xMessage",
+          "name": "startPointer",
+          "type": "xMessage",
+          "label": "start Pointer"
+        }
+      ]
+    },
+    "type": "customNode",
+    "width": 300,
+    "height": 113,
+    "dragging": false,
+    "position": {
+      "x": 409.5560193025037,
+      "y": 25.92199759211908
+    },
+    "selected": false,
+    "positionAbsolute": {
+      "x": 409.5560193025037,
+      "y": 25.92199759211908
+    },
+    "style": {}
+}
