+++ compiler/types.js
@@ -1,0 +1,14 @@
+const INPUT_FIELD_TYPES = {
+    INPUT: "Input",
+    SELECT: "Select",
+    TEXTAREA: "Input.TextArea",
+};
+
+const VALIDATION_TYPES = {
+    URL: "url",
+    EMAIL: "email",
+    DATE: "date",
+    NUMBER: "number",
+};
+
+module.exports = {INPUT_FIELD_TYPES, VALIDATION_TYPES};
